Increase network setup timeout for Windows and protect cleanupNS for concurrent calls.
